Cleanup of unnecessary files, funcs and typos. Adds websocket test html file
