+++ UI_functions.py
@@ -1,5 +1,3 @@
-# Authors: Matteo Golin
-# Imports
 import queue
 import packets
